[kernel] 将涉及到内核libc的部分转移到klibc.c文件中
